requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
python-dotenv==1.0.1
Authlib==1.3.1
itsdangerous>=2.2.0
httpx>=0.27.0
